Fix the type of variable for blocknr

+++ aeon_extents.h
@@ -38,7 +38,8 @@ struct aeon_extent *aeon_search_extent(struct super_block *sb,
 				       struct aeon_inode_info_header *sih,
 				       unsigned long iblock);
 int aeon_update_extent(struct super_block *sb, struct inode *inode,
-		       unsigned blocknr, unsigned long offset, int num_blocks);
+		       unsigned long blocknr, unsigned long offset,
+		       int num_blocks);
 int aeon_rebuild_rb_extenttree(struct super_block *sb,
 			       struct inode *inode, int entries);
 #endif
